requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.0
trafilatura==1.8.0
lxml==5.3.0
pydantic==2.8.2
python-dotenv==1.0.1
google-generativeai==0.7.2
tenacity==9.0.0
tqdm==4.66.4